# docs/swagger.yaml
openapi: 3.0.2
info:
  title: Barbers-API
  version: 1.0.0
  description: |
    ### This is a Swagger UI for our Barbers API
paths:
...
  '/barbers/{id}':
    get:
      tags:
        - details
      parameters:
        - examples:
            ID:
              value: '65'
          name: id
          description: ID of barber.
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              examples:
                Details of a barber:
                  value: |
                    { id: 1, name: "John Wick", working_day: "Monday", specialization: "Haircuts" }
          description: Return full details of a given id barber
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Game/w+ was not found:
                  value:
                    error: barber \w+ was not found in the game array.
          description: Game was not found in the games array.
      operationId: id
      summary: Returns full details of a given barber.
      description: Returns full details of a given barber, based on user input id by the user.
...
  '/bookings/{id}':
    get:
      tags:
        - bookings
      parameters:
        - name: id
          in: path
          description: ID of booking.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Return full details of a given id booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bookingDetails'
        '404':
          description: Booking was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
    error:
      title: Root Type for error
      description: ''
      type: object
      properties:
        error:
          type: string
      example:
        error: Error message.
...
    bookingDetails:
      title: Root Type for booking details
      description: Booking details object
      type: object
      properties:
        id_booking:
          type: integer
        bookingDate:
          type: string
          format: date
        id_customer:
          type: integer
        id_barber:
          type: integer
...